refactor(request): extract merge inflight tracking
